+++ bin/github-publish.js
@@ -3,8 +3,8 @@
 
 var dotenv = require('dotenv');
 var dateFns = require('date-fns');
-var prepareStats = require('./prepare-stats-Cw77weSP.js');
-var fsUtils = require('./fs-utils-BE2k8eQv.js');
+var prepareStats = require('./prepare-stats-DBSeIwsz.js');
+var fsUtils = require('./fs-utils-quNc_MH0.js');
 var webApi = require('@slack/web-api');
 require('@octokit/core');
 require('node:fs');