+++ bin/github-stats.js
@@ -2,9 +2,9 @@
 'use strict';
 
 var dotenv = require('dotenv');
-var prepareStats = require('./prepare-stats-Cw77weSP.js');
+var prepareStats = require('./prepare-stats-DBSeIwsz.js');
 var dateFns = require('date-fns');
-require('./fs-utils-BE2k8eQv.js');
+require('./fs-utils-quNc_MH0.js');
 require('@octokit/core');
 require('node:fs');
 require('node:fs/promises');